New Version: 1.1.0, New License: BSD
The library is now licensed under the Simplified BSD License, allowing
it to be used for commercial projects!

The library is now in version 1.1.0 due to the incompatibility changes
and the new license.

Documentation has been built for the c4dtools.misc module.

Changes:

- Added function c4dtools.utils.update_editor()
- Added function c4dtools.utils.get_shader_bitmap()
- Changed implementation of c4dtools.uitls.Importer.import_. It now
makes use of the built-in __import__() function and does not rely on the
imp module anymore. The paths in the system-path are appended to the
paths defined in the importer.
- c4dtools.prepare() now requires a second argument at minimum: The
plugin's __res__ (c4d.plugins.GeResource) object. This is necessary for
the new string-loading mechanism
- c4dtools.resource.Resource now requires a c4d.plugins.GeResource
object on construction due to the new string-loading mechanism.
- c4dtools.resource.StringLoader.__getattribute__() now returns
c4dtools.resource.ResourceString objects. It is compatible with the
callable object that was returned in previous versions.
- c4dtools.resource.ResourceString class has been added. It combines a
resource-id and a c4d.plugins.GeResource object and implements mechanism
for loading the associated string (via __call__()) and more (such as
retrieving a tuple of the id and the string, see ResourceString.both)